fixed smart completion of tags
